pyproject.toml
[project]
name = "codeapi"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "openai>=2.6.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "tortoise-orm>=0.25.1",
    "uvicorn[standard]>=0.38.0",
]
